backend/requirements.txt
fastapi
uvicorn[standard]
ollama
nemoguardrails
openai-whisper
suno-bark
scipy
torch
